refactor: move price functions from solar_utils to price_calculator
Extract price calculation functions from solar_utils.py to price_calculator.py
for better module organization:

- Move get_price_for_slot() to price_calculator.py
- Move get_price_for_slot_or_none() to price_calculator.py
- Update all imports in dependent modules:
  - excess_solar.py
  - fit_analyzer.py
  - forecast_computer.py
  - proactive_export.py
  - soc_simulator.py
- Update __init__.py exports

This improves separation of concerns - solar_utils now focuses solely on
solar forecasting, while price_calculator handles Amber price lookups.

Closes #149
